refactor: customLib.js 파일 위치 이동 및 경로 정리
